Fix panics and unsafe code
This change fixes panics (#6) and unsafe code (#5). This comes at the
cost of an additional copy of the data send through the pipe and having
a buffer in the state.

Moreover all unsafe code is removed and the need for a custom `Drop`
implementation which makes the code overall easier.

We also add tests.